requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
ExifRead==3.0.0
idna==3.3
pycryptodome==3.14.1
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
watchdog==2.1.8
